+++ clock.go
@@ -1,0 +1,11 @@
+package hhru
+
+import "time"
+
+type Clock interface {
+	Now() time.Time
+}
+
+type wallClock struct{}
+
+func (wallClock) Now() time.Time { return time.Now() }